DS-67 Moved fudis-checkbox under forms
