+++ ngx-fudis/projects/ngx-fudis/src/public-api.ts
@@ -8,7 +8,7 @@ export * from './lib/components/dialog/public-api';
 export { HeadingComponent } from './lib/components/typography/heading/heading.component';
 export { BodyTextComponent } from './lib/components/typography/body-text/body-text.component';
 export { ButtonComponent } from './lib/components/button/button.component';
-export { FudisCheckboxComponent } from './lib/components/fudis-checkbox/fudis-checkbox.component';
+export { FudisCheckboxComponent } from './lib/components/form/fudis-checkbox/fudis-checkbox.component';
 export { CheckboxComponent } from './lib/components/form/checkbox/checkbox.component';
 export { TextInputComponent } from './lib/components/form/text-input/text-input.component';
 export { ErrorSummaryComponent } from './lib/components/form/error-summary/error-summary.component';
